project_plan_generator/requirements.txt
pydantic-ai
python-dotenv
openai
pydantic
